+++ packages/lodestar/src/opPool/opPool.ts
@@ -21,7 +21,7 @@ import {
 import {IEth1Notifier} from "../eth1";
 import {IBeaconConfig} from "@chainsafe/eth2.0-config";
 import {computeEpochOfSlot} from "../chain/stateTransition/util";
-import { blockToHeader } from "../chain/stateTransition/util/block";
+import { blockToHeader } from "../chain/stateTransition/util/blockRoot";
 
 interface IOpPoolModules {
   config: IBeaconConfig;
